Unit test fix 2nd time
